requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
httplib2==0.19.1
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
msgpack==1.0.2
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
proto-plus==1.18.1
protobuf==3.15.7
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
